+++ src/opi/output/grepper/recipes.py
@@ -5,6 +5,7 @@ from opi.output.grepper.patterns import (
     CASSCF_CONVERGED,
     CC_CONVERGED,
     ERROR_PATTERNS,
+    EXTERNAL_XTB_SCF_CONVERGED,
     GEOMETRY_CONVERGED,
     HAS_ABORTING,
     HAS_FREQ,
@@ -280,7 +281,7 @@ def has_scf(file_name: Path, /) -> bool:
 
 def has_scf_converged(file_name: Path, /) -> bool:
     """
-    Searches for the message 'SUCCESS' as indicator that the SCF converged.
+    Searches for the message 'SUCCESS' or 'convergence criteria satisfied after' as indicator that the SCF converged.
 
     Parameter
     ---------
@@ -292,7 +293,9 @@ def has_scf_converged(file_name: Path, /) -> bool:
     bool
         True if expression is found in file else False
     """
-    return has_string_in_file(file_name, SCF_CONVERGED)
+    return has_string_in_file(file_name, SCF_CONVERGED) or has_string_in_file(
+        file_name, EXTERNAL_XTB_SCF_CONVERGED
+    )
 
 
 def has_casscf_converged(file_name: Path, /) -> bool: